both pyarrow and fastparquet working, but using pd.concat
