+++ dask/dataframe/io/parquet/utils.py
@@ -67,26 +67,6 @@ class Engine:
         raise NotImplementedError()
 
     @staticmethod
-    def aggregate_row_groups(parts, stats, batch_size, **kwargs):
-        """
-        Aggregate row-groups, and produce new parts and stats objects.
-
-        Parameters
-        ----------
-        parts: List
-            Contains metadata objects to write, of the type undrestood by
-            the specific implementation.
-        statistics: List[Dict]
-            A list of dictionaries of statistics data, one dict for every
-            partition.
-        batch_size: int
-            The ideal number of rows to use for a single task/paritition.
-        **kwargs: dict
-            Other keyword arguments
-        """
-        raise NotImplementedError()
-
-    @staticmethod
     def read_partition(fs, piece, columns, index, **kwargs):
         """ Read a single piece of a Parquet dataset into a Pandas DataFrame
